requirements.txt
protobuf~=6.31.1
google-auth-oauthlib~=1.2.2
google-api-python-client~=2.171.0